+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [1.1] - 2024-07-08
+### Changed
+- `JSONPointer`: added `from(string)`
+- `JSONPointerException`: made class open and rearranged properties to allow for extension
+- `pom.xml`: updated dependency version
+
 ## [1.0] - 2024-07-04
 ### Added
 - all files: initial versions